there should only be one `<slot name="description-label">` for each `<verifiable-credential>`
if there is a `<slot name="description-label">` for each item in `"output_descriptors"` then only the first one will actually get slotted as the `name` will clash (just like `id`)

also, the nature of `<slot>` means that the slotted content will be used as-is instead of matching the styling of the `<verifiable-credential>` (e.g. it could have red text)

instead, have a single top-level `<slot name="description-label">` that updates every "Description" label whenever something is slotted or if the slotted node changes (e.g. `textContent`) using the `textContent` so that no styles are copied/overridden